+++ r/R/zzz.R
@@ -1,8 +1,13 @@
 pkg <- new.env(parent = emptyenv())
 
 .onLoad <- function(libname, pkgname) {
-  pkg$catalog_url <- "https://data.bedrock.bio/manifest.json"
-  pkg$credentials_url <- "https://data.bedrock.bio/credentials.json"
+  host <- if (identical(Sys.getenv("BB_ENV"), "dev")) {
+    "https://data-dev.bedrock.bio"
+  } else {
+    "https://data.bedrock.bio"
+  }
+  pkg$manifest_url <- paste0(host, "/manifest.json")
+  pkg$credentials_url <- paste0(host, "/credentials.json")
 }
 
 .onUnload <- function(libpath) {